requirements.txt
torch>=2.2
torchvision>=0.17
transformers>=4.40
pandas>=2.0
pillow>=10.0
matplotlib>=3.8
tqdm>=4.66
scikit-learn>=1.4